requirements.txt
beautifulsoup4
celery
face_recognition
numpy
pandas
psycopg2
requests
selenium
selenium-screenshot